src/requirements.txt
aiohttp==3.6.2
aiohttp-json-rpc==0.13.2
aiosqlite==0.15.0
appdirs==1.4.4
apsw-wheels==3.30.1.post3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
ChatterBot==1.1.0
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
discord.py==1.5.1
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
future==0.18.2
fuzzywuzzy==0.18.0
idna==2.10
jieba3k==0.35.1
joblib==0.17.0
kiwisolver==1.3.1
lxml==4.6.1
Markdown==3.2.2
matplotlib==3.3.3
multidict==4.7.6
murmurhash==1.0.4
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.1
pytz==2020.1
PyYAML==5.3.1
Red-DiscordBot==3.4.3
Red-Lavalink==0.7.1
regex==2020.11.13
requests==2.25.0
requests-file==1.5.1
sacremoses==0.0.43
schema==0.7.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
sentencepiece==0.1.91
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.0.2
tokenizers==0.9.3
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.48.0
transformers==3.5.1
typing-extensions==3.7.4.2
urllib3==1.26.2
uvloop==0.14.0
wasabi==0.8.0
yarl==1.5.1
